rename ReferenceSearch to NoteParser
